Moved utils, fixed scenario ref
